CompositeBuildings

Documentation for CompositeBuildings.

This package provides functions to load various datasets containing building data into consistents DataFrames.DataFrames.

It as well provides the functionality to calculate shadows from these datasets.

Interface

To properly work within the MinistryOfCoolWalks ecosystem, we expect the dataframes to follow a certain set of requirements:

Each row in the DataFrame represents a single building

Each DataFrame must have the following columns:

:id unique id of the building

:height height of building (m)

:geometry ArchGDAL polygon (single polygon, possibly with holes), with spatial reference applied.

The DataFrame must have metadata with a key of observatory, which contains a CoolWalksUtils.ShadowObservatory. This value contains the center coordinates used for projection to a local coordinate system and the timezone of the dataset, used to calculate the sunposition, for shadow projection.

The technical part of the last two of these requirements can be checked with CompositeBuildings.check_building_dataframe_integrity(df).

API

CompositeBuildings.check_building_dataframe_integrity

CompositeBuildings.convex_report

CompositeBuildings.check_building_dataframe_integrity — Method

check_building_dataframe_integrity(df)

Checks if df conforms to the technical requirements needed to be considered as a source for building data.

CompositeBuildings.convex_report — Method

convex_report(df)

Little utility to log how many of the entries indf.geometry` are convex.