Other Building Loaders

Introduction

Here we compile a list of convenience functions which load building footprints from different sources and prepare them for shadow casting.

API

CompositeBuildings.load_british_shapefiles

CompositeBuildings.load_new_york_shapefiles

CompositeBuildings.load_british_shapefiles — Method

load_british_shapefiles(path; bbox=nothing)

loads the shapefiles of the largest cities in great britain, provided by emu analytics into dataframes, possibly clipping along a named tuple BoundingBox with names (minlon, minlat, maxlon, maxlat).

Returns a dataframe with the columns given in the shapefile, with a few exceptions: :OBJECTID => :id, :MEAN_mean => :height_mean, :MIN_min => :height_min, :MAX_max => :height_max. Polygons are stored in the geometry column in EPSG 4326 crs. Only rows where :height_mean>0 are returned.

The dataframe has metadata of center_lat and center_lon, representing the central latitude and longitude of the bounding Box, applied.

CompositeBuildings.load_new_york_shapefiles — Method

load_new_york_shapefiles(path; bbox=nothing)

loads the shapefiles containing building footprints and heights provided by the city of new york into dataframes, possibly clipping along a named tuple BoundingBox with names (minlon, minlat, maxlon, maxlat).

Returns a dataframe with the columns given in the shapefile, only rows where heightroof is not missing and larger than 0 are returned. Polygons are stored in the geometry column in EPSG 4326 crs.

The dataframe has metadata of center_lat and center_lon, representing the central latitude and longitude of the bounding Box, applied.