CompositeBuildings

Documentation for CompositeBuildings.

CompositeBuildings.BuildingPart — Type

OSM Building Part, consisting out of one ArchGDAL Polygon

CompositeBuildings.CompositeBuilding — Type

Composite Building, consisting of a SimpleBuilding (as a fallback) and a vector of building parts

CompositeBuildings.SimpleBuilding — Type

Simple OSM Building, consisting out of one ArchGDAL Polygon

CompositeBuildings.relate_buildings — Function

constructs a dataframe which contains two id columns, describing the relation between two dataframes. Both dataframes are expected to have at least an :id and an :geometry column. Two geometries g1, g2 are related, if: area(intersection(g1, g2)) >= min(area(g1), area(g2)) * overlap