Other Building Loaders

Introduction

Here we compile a list of convenience functions which load building footprints from different sources and prepare them for shadow casting.

The returned DataFrames conform to the requirements needed to be a source for building data.

API

CompositeBuildings.load_british_shapefiles

CompositeBuildings.load_new_york_shapefiles

CompositeBuildings.load_british_shapefiles — Method

load_british_shapefiles(path; extent=nothing)

loads the shapefiles of the largest cities in great britain, provided by emu analytics into a DataFrame.

arguments

path: Path to the file with the dataset

extent: Extents.Extent, specifying a clipping range for the Dataset. Use X for lon and Y for lat.

Returns a dataframe with the columns given in the shapefile, with a few exceptions:

:OBJECTID => :id

:MEAN_mean => :height

:MIN_min => :height_min

:MAX_max => :height_max

Polygons are stored in the geometry column in EPSG 4326 crs. Only rows where :height > 0 are returned.

The dataframe has a metadata tag observatory, containing the relevant center.

CompositeBuildings.load_new_york_shapefiles — Method

load_new_york_shapefiles(path; extent=nothing)

loads the shapefiles containing building footprints and heights provided by the city of new york into a DataFrame.

arguments

path: Path to the file with the dataset

extent: Extents.Extent, specifying a clipping range for the Dataset. Use X for lon and Y for lat.

Returns a dataframe with the columns given in the shapefile, only rows where heightroof is not missing and larger than 0 are returned. Polygons are stored in the geometry column in EPSG 4326 crs.

The dataframe conforms to the requirements needed to be a source for buildings.