Spain Building Loaders and Utilities

Introduction

Spain provides easy access to their cadastral dataset of building footprints and metadata. We provide some functions to inspect, download, preprocess and load the available data.

The returned DataFrames conform to the requirements needed to be a source for building data.

Usage example

API

CompositeBuildings.download_spain_overview

CompositeBuildings.download_spain_region_overview

CompositeBuildings.download_spain_subregion

CompositeBuildings.load_spain_buildings_gml

CompositeBuildings.load_spain_parts_gml

CompositeBuildings.load_spain_processed_buildings

CompositeBuildings.parse_polygon_string

CompositeBuildings.parse_spain_xml

CompositeBuildings.preprocess_spain_subregion

CompositeBuildings.relate_floors

Inspect datasets

CompositeBuildings.parse_polygon_string — Function

parse_polygon_string(polystring)

parses a string of the form "lat1 lon1 lat2 lon2..." or "lat1\tlon1\tlat2\tlon2..." into an ArchGDAL polygon.

CompositeBuildings.parse_spain_xml — Function

parse_spain_xml(url)

Reads an XML Document from the Spanish Cadastral Dataset (see here for an example) and parses it into a dataframe containing information about the regions described within it.

CompositeBuildings.download_spain_overview — Function

download_spain_overview()

Downloads and parses the xml file at https://www.catastro.minhap.es/INSPIRE/buildings/ES.SDGC.BU.atom.xml

CompositeBuildings.download_spain_region_overview — Function

download_spain_region_overview(region_id, spain_overview=download_spain_overview())

Downloads and parses the xml corresponding to the region_id from the cadastral dataset. Can optionally take the result from download_spain_overview(), to reduce the number of server calls.

Download and process

CompositeBuildings.download_spain_subregion — Function

download_spain_subregion(url, savepath)

Downloads the zip file at url (element of download_spain_region_overview(...).url) to savepath.zip, extracts it to savepath/raw, and deletes the zip file.

CompositeBuildings.load_spain_buildings_gml — Function

load_spain_buildings_gml(path; extent=nothing)

Loads the file at path with GeoDataFrames, and applies some transformations that only make sense if it is a buildings file from the spanish cadastral dataset.

CompositeBuildings.load_spain_parts_gml — Function

load_spain_parts_gml(path; extent=nothing)

Loads the file at path with GeoDataFrames, and applies some transformations that only make sense if it is a buildingparts file from the spanish cadastral dataset.

CompositeBuildings.relate_floors — Function

relate_floors(buildings, buildings_parts)

Infers the number of floors a building in buildings might have, given the data present in its constituent buildings_parts by various different methods:

nFloors_overlap: average number of floors in all parts, weighed by intersection area (overlap) between each part and the building.

nFloors_approx: given area in the dataset divided by area of the footprint (myArea). Assumes that area is a "usable area".

nFloors_part_approx: given area in the dataset divided by the total footprint-area of all parts (myArea_part). Assumes that the given area is some kind of "usable area".

maxFloors_part: maximum value of floors in all parts associated with building.

CompositeBuildings.preprocess_spain_subregion — Function

preprocess_spain_subregion(path; extent=nothing)

Loads buildings and building parts form path/raw, and save the result of relate_floors(...) to path/buildings.geojson

Load

CompositeBuildings.load_spain_processed_buildings — Function

load_spain_processed_buildings(path; extent=nothing, floor_height=4.0)

loads the files saved by preprocess_spain_subregion(...) from path/buildings.geojson and adds metadata for projection.

Uses floor_height to convert nFloor_overlap to building height.

The returned DataFrame fulfills the requirements to be used as a source for building data.